Implement program to print C variable declarations
This C program prints all variable declarations inside a C source file.
Each declaration is preceded by the line number where it's located in
the source file. Output to the terminal uses ANSI colors.

Note: This program (currently) does NOT display structure or union
declarations. It only prints variable declarations which are of a
structure or union type.